Add multi-browser Geb config based on + sample test
Based on https://github.com/bonigarcia/webdrivermanager and the
corresponding GebConfig we can now test with these browsers by just
changing Maven property geb.env:
  - Chrome
  - Firefox
  - IE
  - Edge
  - PhantomJS
  - HTMLUnit (incl. JavaScript!)
  - Opera (theoretically, if Opera Chromium Driver was actively
    maintained and functional in current versions)

SampleGebIT is actually unrelated to this project (as is Geb in general,
but I wanted to commit & push it somewhere). It shows how to shut down
and restart a Selenium driver within a running one Geb spec. See
http://stackoverflow.com/questions/42069291.